feat: paginate historical price EOD endpoints past the 5000-record API limit

+++ src/Mapper/HistoricalPriceEodNonSplitAdjustedMapper.php
@@ -33,7 +33,7 @@ final readonly class HistoricalPriceEodNonSplitAdjustedMapper extends Type
 			'adjHigh' => $ts->float(),
 			'adjLow' => $ts->float(),
 			'adjClose' => $ts->float(),
-			'volume' => $ts->int(),
+			'volume' => $ts->float(),
 		],
 			identifier: 'symbol',
 		);
